Refactor internet radio loading into a service
